+++ docs_src/src/guide/changelog.md
@@ -11,3 +11,4 @@
 - Added [request data accessors](./basics/requests#accessors).
 - Some refactoring and package renaming have been done to better respect the Go conventions.
     - The `helpers` package have been renamed to `helper`
+- The server now shuts down when it encounters an error.